git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/branches/prism-games@7995 bbc10eb1-c90d-0410-af57-cb519fbb1720

+++ prism/src/prism/PrismCL.java
@@ -183,7 +183,7 @@ public class PrismCL implements PrismModelListener
 	private double simWidth;
 	private int reqIterToConclude;
 	private double simMaxReward;
-	private int simMaxPath;
+	private long simMaxPath;
 	private boolean simApproxGiven = false;
 	private boolean simConfidenceGiven = false;
 	private boolean simNumSamplesGiven = false;
@@ -280,7 +280,7 @@ public class PrismCL implements PrismModelListener
 			if (simpath) {
 				try {
 					if (!simMaxPathGiven)
-						simMaxPath = prism.getSettings().getInteger(PrismSettings.SIMULATOR_DEFAULT_MAX_PATH);
+						simMaxPath = prism.getSettings().getLong(PrismSettings.SIMULATOR_DEFAULT_MAX_PATH);
 					File f = (simpathFilename.equals("stdout")) ? null : new File(simpathFilename);
 					prism.generateSimulationPath(modulesFile, simpathDetails, simMaxPath, f);
 				} catch (PrismException e) {
@@ -1550,7 +1550,7 @@ public class PrismCL implements PrismModelListener
 				else if (sw.equals("simpathlen")) {
 					if (i < args.length - 1) {
 						try {
-							simMaxPath = Integer.parseInt(args[++i]);
+							simMaxPath = Long.parseLong(args[++i]);
 							if (simMaxPath <= 0)
 								throw new NumberFormatException("");
 							simMaxPathGiven = true;
@@ -2036,7 +2036,7 @@ public class PrismCL implements PrismModelListener
 		if (!simMaxRewardGiven)
 			simMaxReward = prism.getSettings().getDouble(PrismSettings.SIMULATOR_MAX_REWARD);
 		if (!simMaxPathGiven)
-			simMaxPath = prism.getSettings().getInteger(PrismSettings.SIMULATOR_DEFAULT_MAX_PATH);
+			simMaxPath = prism.getSettings().getLong(PrismSettings.SIMULATOR_DEFAULT_MAX_PATH);
 
 		// Pick a default method, if not specified
 		// (CI for quantitative, SPRT for bounded)
